Split KotlinWhenExpressionTarget
